recordeca_gmtk_jam_2025/requirements.txt
linux-joystick-py
pynput
